refactor(alchemy-select): extract shared Tom Select settings helper
The Floating UI dropdown positioning and the i18n "create"/"no results"
messages lived inline in the alchemy-select component. The upcoming tags
autocomplete needs the same behavior but cannot inherit from that
<select>-based component, so move the shared pieces into a side-effect-free
utils/tom_select module that both can import. This keeps the base component
focused and makes the shared configuration independently testable.
